add error

+++ src/app/app.js
@@ -33,7 +33,11 @@ app.get('/api/v1/users/:id', (req, res) => {
   const id = req.params.id
 
   db.get(`SELECT * FROM users WHERE id = ${id}`, (err, row) => {
-    res.json(row)
+    if (!row) {
+      res.status(404).send({error:"Not Found"})
+    } else {
+      res.json(row)
+    }
   })
 
   db.close()
@@ -53,41 +57,58 @@ app.get('/api/v1/search', (req, res) => {
 
 // Create a new user
 app.post('/api/v1/users', async (req, res) => {
-  const db = new sqlite3.Database(dbPath)
+  if (!req.body.name || !req.body.name === "") {
+    res.status(400).send({error:"ユーザー名が指定されていません"})
+  } else {
+    const db = new sqlite3.Database(dbPath)
 
-  const name = req.body.name
-  const profile = req.body.profile ? req.body.profile : ""
-  const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
+    const name = req.body.name
+    const profile = req.body.profile ? req.body.profile : ""
+    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : ""
 
-  await run(`INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}", "${profile}", "${dateOfBirth}")`,
-            db,
-            res,
-            "Create Success!"
-            )
+    try {
+      await run(`INSERT INTO users (name, profile, date_of_birth) VALUES ("${name}", "${profile}", "${dateOfBirth}")`,
+      db
+      )
+      res.status(201).send({message: "Create!"})
+    } catch (e) {
+      res.status(500).send({error: e})
+    }
 
-  db.close()
+    db.close()
+  }
 })
 
 // Update user
 app.put('/api/v1/users/:id', async (req, res) => {
-  const db = new sqlite3.Database(dbPath)
+  if (!req.body.name || !req.body.name === "") {
+    res.status(400).send({error:"ユーザー名が指定されていません"})
+  } else {
+    const db = new sqlite3.Database(dbPath)
 
-  const id = req.params.id
+    const id = req.params.id
 
-  db.get(`SELECT * FROM users WHERE id = ${id}`, async (err, row) => {
-    const name = req.body.name ? req.body.name : row.name
-    const profile = req.body.profile ? req.body.profile : row.profile
-    const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
+    db.get(`SELECT * FROM users WHERE id = ${id}`, async (err, row) => {
 
-    console.log(`UPDATE users SET name="${name}", profile="${profile}", date_of_birth="${dateOfBirth}" WHERE id=${id}`)
-    await run(`UPDATE users SET name="${name}", profile="${profile}", date_of_birth="${dateOfBirth}" WHERE id=${id}`,
-              db,
-              res,
-              "Update Success!"
-    )
-  })
+      if (!row) {
+        res.status(404).send({error:"Not Found"})
+      } else {
+        const name = req.body.name ? req.body.name : row.name
+        const profile = req.body.profile ? req.body.profile : row.profile
+        const dateOfBirth = req.body.date_of_birth ? req.body.date_of_birth : row.date_of_birth
 
-  db.close()
+        try {
+          await run(`UPDATE users SET name="${name}", profile="${profile}", date_of_birth="${dateOfBirth}" WHERE id=${id}`,
+                    db
+          )
+          res.status(200).send({message: "Update!"})
+        } catch (e) {
+          res.status(500).send({error: e})
+        }
+      }
+    })
+    db.close()
+  }
 })
 
 // Delete user
@@ -95,24 +116,32 @@ app.delete('/api/v1/users/:id', async (req, res) => {
   const db = new sqlite3.Database(dbPath)
   const id = req.params.id
 
-  await run(`DELETE FROM users WHERE id=${id}`,
-            db,
-            res,
-            "Delete Success!"
-  )
+  db.get(`SELECT * FROM users WHERE id = ${id}`, async (err, row) => {
+
+    if (!row) {
+      res.status(404).send({error:"Not Found"})
+    } else {
+      try {
+        await run(`DELETE FROM users WHERE id=${id}`,
+                  db
+        )
+        res.status(200).send({message:"Delete!"})
+      } catch(e) {
+        res.status(500).send({error: e})
+      }
+    }
 
   db.close()
+  })
 })
 
 
-const run = async(sql, db, res, message) => {
+const run = async(sql, db) => {
   return new Promise((resolve, reject) => {
     db.run(sql, (err) => {
       if (err) {
-        res.status(500).send(err)
-        return reject()
+        return reject(err)
       } else {
-        res.json({message: message})
         return resolve()
       }
     })
